requirements.txt
flask==2.3.3
sqlalchemy==2.0.20
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
plotly==5.16.1
dash==2.13.0
wtforms==3.0.1
flask-wtf==1.1.1
flask-login==0.6.2
pytest==7.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
flask-sqlalchemy==3.0.5
pymongo==4.5.0
nltk==3.8.1
transformers==4.33.1
flask-cors==4.0.0
jsonschema==4.19.0